mant. empresas a medias

+++ server/src/controller/empresas.controller.js
@@ -50,10 +50,13 @@ export const updateEmpresaById = async (req, res) => {
 };
 
 export const deleteEmpresaById = async (req, res) => {
+    console.log("entro delete")
+    console.log("1 params:" + req.params.empresaId)
     try {
-        const {empresaIdDelete} = req.params
+        const empresaIdDelete = await empresas.findOne({_id: req.params.empresaId});
+        if(!empresaIdDelete) return res.status(400).json({message: "la empresa no existe"})
         await empresas.findByIdAndDelete(empresaIdDelete)
-        res.status(204).json
+        res.status(204).json({msg: "empresa borrada con éxito"});
     } catch (error) {
         console.log(error)
         res.status(400).json({msg: "error borrando empresa por ID"});
@@ -64,6 +67,7 @@ export const getEmpresaByName = async (req, res) => {
     try {
         const nombreEmpresa = await empresas.findOne({nombre: req.params.empresaName});
         if(!nombreEmpresa) return res.status(400).json({message: "la empresa no existe"})
+        console.log(nombreEmpresa)
         res.status(200).json({nombreEmpresa, message: "empresa encontrada"})
     } catch (error) {
         console.log(error)
